CyborgDB/requirements.txt
Flask
flask-cors
cryptography
cyborgdb
numpy
protobuf
python-dotenv
redis
requests
torch
transformers
google-genai
gunicorn
